@@ -5,11 +5,12 @@
   "main": "index.js",
   "scripts": {
     "lint": "eslint src --ext .js --ext .jsx",
-    "start": "webpack-dev-server --progress --colors --display-chunks --display-modules --display-origins",
-    "build": "npm run lint; NODE_ENV=production webpack --optimize-minimize --progress --colors",
-    "dev-build": "npm run lint; webpack --progress --colors",
+    "start": "webpack-dev-server --progress --colors",
+    "build": "NODE_ENV=production webpack --optimize-minimize --progress --colors",
+    "dev-build": "webpack --progress --colors",
     "start-webdriver": "webdriver-manager update && webdriver-manager start",
     "test:only": "ava --verbose --serial",
+    "pretest": "npm run lint",
     "test": "./script/run-test.sh",
     "deploy": "./script/deploy.sh"
   },
@@ -24,24 +25,26 @@
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-react": "^6.3.13",
     "css-loader": "^0.27.0",
-    "eslint": "^3.8.1",
+    "eslint": "^4.1.1",
+    "eslint-import-resolver-webpack": "^0.8.3",
     "eslint-loader": "^1.1.1",
-    "eslint-plugin-react": "^6.4.1",
-    "extract-text-webpack-plugin": "^1.0.1",
+    "eslint-plugin-import": "^2.6.0",
+    "eslint-plugin-react": "^7.1.0",
+    "extract-text-webpack-plugin": "^2.1.2",
     "favico.js": "^0.3.10",
     "file-loader": "^0.11.0",
-    "http-server": "^0.9.0",
+    "http-server": "^0.10.0",
     "json-loader": "^0.5.4",
     "less": "^2.5.3",
     "less-loader": "^3.0.0",
     "node-libs-browser": "^2.0.0",
-    "parallelshell": "^2.0.0",
+    "parallelshell": "^3.0.1",
     "react-hot-loader": "^3.0.0-beta.6",
     "url-loader": "^0.5.6",
-    "webdriver-manager": "^11.1.0",
+    "webdriver-manager": "^12.0.6",
     "webdriverio": "^4.5.2",
-    "webpack": "^1.9.10",
-    "webpack-dev-server": "^1.9.0"
+    "webpack": "^3.0.0",
+    "webpack-dev-server": "^2.5.0"
   },
   "dependencies": {
     "babel-polyfill": "^6.3.14",
@@ -49,30 +52,33 @@
     "c3": "^0.4.11",
     "classnames": "^2.2.1",
     "d3": "^4.3.0",
-    "dexie": "^1.3.1",
+    "dexie": "2.0.0-beta.7",
+    "eslint-plugin-duck": "0.5.6",
     "etherpad-lite-client": "philschatz/etherpad-lite-client-js#master",
     "gfm-linkify": "^0.1.0",
-    "history": "^4.5.0",
-    "jsonquery-engine": "^1.0.2",
-    "level-js": "philschatz/level.js#master",
-    "level-queryengine": "^3.0.1",
-    "levelidb": "^0.2.6",
-    "levelup": "^1.3.1",
-    "memdown": "^1.1.2",
-    "mermaid": "^7.0.0",
-    "moment": "2.18.0",
+    "history": "^3.2.1",
+    "mermaid": "^7.0.3",
+    "moment": "^2.18.1",
     "octicons": "^5.0.1",
     "octokat": "^0.6.2",
     "puzzle-script": "0.0.2",
     "qs": "^6.3.0",
     "react": "^15.3.2",
-    "react-bootstrap": "^0.30.5",
+    "react-bootstrap": "^0.31.0",
     "react-dnd": "^2.1.2",
     "react-dnd-html5-backend": "^2.1.2",
     "react-dom": "^15.3.2",
-    "react-octicons": "0.0.5",
+    "react-octicons": "0.0.4",
+    "react-redux": "^5.0.2",
     "react-router": "^3.0.0",
+    "react-router-redux": "^4.0.7",
     "recordo": "^0.0.6",
+    "reduck": "0.2.1",
+    "redux": "^3.6.0",
+    "redux-logger": "^3.0.6",
+    "redux-optimist": "^0.0.2",
+    "redux-optimist-promise": "^0.8.0",
+    "redux-toast-feedback": "^0.1.0",
     "ultramarked": "^1.7.0",
     "underscore": "^1.8.3"
   }
