@@ -1,18 +1,17 @@
 import React from 'react';
+import {connect} from 'react-redux';
 import _ from 'underscore';
-import * as BS from 'react-bootstrap';
 import {Link} from 'react-router';
 import {MilestoneIcon} from 'react-octicons';
 
 import {getFilters} from '../route-utils';
-import IssueStore from '../issue-store';
-import SettingsStore from '../settings-store';
-import FilterStore from '../filter-store';
-import CurrentUserStore from '../user-store';
-import Loadable from './loadable';
+import {getReposFromStr} from '../helpers';
+import {
+  fetchMilestones,
+  fetchIssues
+} from '../redux/ducks/issue';
 import IssueList from './issue-list';
 import Issue from './issue';
-import Board from './board';
 import GithubFlavoredMarkdown from './gfm';
 
 
@@ -27,7 +26,7 @@ const KanbanColumn = React.createClass({
           primaryRepoName={primaryRepoName}
           card={card}
           columnRegExp={columnRegExp}
-          />
+        />
       );
     });
 
@@ -46,119 +45,124 @@ const KanbanColumn = React.createClass({
       heading = 'No Milestone';
     }
 
-    const isShowingColumn = (
-         (!milestone && !SettingsStore.getHideUncategorized())
-      || (milestone && (
-             getFilters().getState().milestoneTitles.length == 0
-          || getFilters().getState().milestoneTitles.indexOf(milestone.title) >= 0))
-    );
-
-    if (isShowingColumn) {
-      return (
-        <div key={milestone && milestone.title || 'no-milestone'} className='kanban-board-column'>
-          <IssueList
-            title={heading}
-            milestone={milestone}
-          >
-            {issueComponents}
-          </IssueList>
-        </div>
-      );
-    } else {
-      return null; // TODO: Maybe the panel should say "No Issues" (but only if it's the only column)
-    }
-
-  }
-});
-
-
-const ByMilestoneView = React.createClass({
-  render() {
-    const {columnData, cards, primaryRepoName, columnRegExp} = this.props;
-
-    const uncategorizedCards = _.filter(cards, (card) => {
-      return !card.issue.milestone;
-    });
-
-    let sortedCards = FilterStore.filterAndSort(cards, true/*isShowingMilestones*/);
-
-    let kanbanColumnCount = 0; // Count the number of actual columns displayed
-    kanbanColumnCount++; // "No Milestone" counts as a column
-
-    const uncategorizedColumn = (
-      <KanbanColumn
-        cards={uncategorizedCards}
-        primaryRepoName={primaryRepoName}
-        columnRegExp={columnRegExp}
-      />
-    );
-
-    const kanbanColumns = _.map(columnData, (milestone) => {
-      // If we are filtering by a kanban column then only show that column
-      // Otherwise show all columns
-      const columnCards = _.filter(sortedCards, (card) => {
-        return card.issue.milestone && card.issue.milestone.title === milestone.title;
-      });
-
-
-      // Show the column when:
-      // isFilteringByColumn = label (the current column we are filtering on)
-      // !isFilteringByColumn && (!getShowEmptyColumns || columnCards.length)
-
-      kanbanColumnCount++; // Count the number of actual columns displayed
-      /*HACK: Column should handle milestones */
-      return (
-        <KanbanColumn
-          key={kanbanColumnCount}
-          milestone={milestone}
-          cards={columnCards}
-          primaryRepoName={primaryRepoName}
-          columnRegExp={columnRegExp}
-        />
-      );
-    });
-
     return (
-      <div className='kanban-board'>
-        {uncategorizedColumn}
-        {kanbanColumns}
+      <div className='kanban-board-column'>
+        <IssueList
+          title={heading}
+          milestone={milestone}
+        >
+          {issueComponents}
+        </IssueList>
       </div>
     );
   }
 });
 
 
-const RepoKanbanShell = React.createClass({
+const ByMilestoneView = React.createClass({
   componentWillMount() {
-    // Needs to be called before `render()`
-    IssueStore.startPolling();
-  },
-  componentWillUnmount() {
-    IssueStore.stopPolling();
-  },
-  renderLoaded() {
-    const {repoInfos, columnRegExp} = getFilters().getState();
+    const {repoInfos, dispatch} = this.props;
     // pull out the primaryRepoName
     const [{repoOwner, repoName}] = repoInfos;
-
-    return (
-      <Board {...this.props}
-        repoInfos={repoInfos}
-        columnRegExp={columnRegExp}
-        type={ByMilestoneView}
-        columnDataPromise={IssueStore.fetchMilestones(repoOwner, repoName)}
-      />
-    );
+    dispatch(fetchIssues(repoInfos));
+    dispatch(fetchMilestones(repoOwner, repoName));
   },
   render() {
+    const {milestones, cards, repoInfos, columnRegExp, settings, filter} = this.props;
+
+    // Get the primary repo
+    const [primaryRepo] = repoInfos;
+
+    const uncategorizedCards = _.filter(cards, (card) => {
+      return !card.issue.milestone;
+    });
+
+    const uncategorizedColumn = (
+      <KanbanColumn
+        settings={settings}
+        cards={uncategorizedCards}
+        primaryRepoName={primaryRepo.repoName}
+        columnRegExp={columnRegExp}
+      />
+    );
+
+    const kanbanColumns = milestones.reduce((prev, milestone) => {
+      if (filter.milestoneTitles.length && filter.milestoneTitles.indexOf(milestone.title) === -1) {
+        return prev;
+      }
+
+      // If we are filtering by a kanban column then only show that column
+      // Otherwise show all columns
+      const columnCards = _.filter(cards, (card) => {
+        return card.issue.milestone && card.issue.milestone.title === milestone.title;
+      });
+
+      /*HACK: Column should handle milestones */
+      prev.push(
+        <KanbanColumn
+          settings={settings}
+          key={milestone.title}
+          milestone={milestone}
+          cards={columnCards}
+          primaryRepoName={primaryRepo.repoName}
+          columnRegExp={columnRegExp}
+        />
+      );
+      return prev;
+    }, []);
 
     return (
-      <Loadable
-        promise={CurrentUserStore.fetchUser()}
-        renderLoaded={this.renderLoaded}
-      />
+      <div className='kanban-board'>
+        {!settings.isHideUncategorized && uncategorizedColumn}
+        {kanbanColumns}
+      </div>
     );
   }
 });
 
-export default RepoKanbanShell;
+export default connect((state, ownProps) => {
+  return {
+    repoInfos: getReposFromStr((ownProps.params || {}).repoStr || ''),
+    settings: state.settings,
+    cards: state.issues.cards,
+    columnRegExp: state.filter.columnRegExp,
+    milestones: state.issues.milestones,
+    filter: state.filter
+  };
+})(ByMilestoneView);
+
+//
+// const RepoKanbanShell = React.createClass({
+//   componentWillMount() {
+//     // Needs to be called before `render()`
+//     IssueStore.startPolling();
+//   },
+//   componentWillUnmount() {
+//     IssueStore.stopPolling();
+//   },
+//   renderLoaded() {
+//     const {repoInfos, columnRegExp} = getFilters().getState();
+//     // pull out the primaryRepoName
+//     const [{repoOwner, repoName}] = repoInfos;
+//
+//     return (
+//       <Board {...this.props}
+//         repoInfos={repoInfos}
+//         columnRegExp={columnRegExp}
+//         type={ByMilestoneView}
+//         columnDataPromise={IssueStore.fetchMilestones(repoOwner, repoName)}
+//       />
+//     );
+//   },
+//   render() {
+//
+//     return (
+//       <Loadable
+//         promise={CurrentUserStore.fetchUser()}
+//         renderLoaded={this.renderLoaded}
+//       />
+//     );
+//   }
+// });
+
+// export default RepoKanbanShell;
