@@ -1,18 +1,14 @@
 import React from 'react';
+import {connect} from 'react-redux';
 import _ from 'underscore';
-import * as BS from 'react-bootstrap';
-import {UserIcon} from 'react-octicons';
+import {PersonIcon} from 'react-octicons';
 
-import {getFilters} from '../route-utils';
-import IssueStore from '../issue-store';
-import FilterStore from '../filter-store';
-import CurrentUserStore from '../user-store';
-import Loadable from './loadable';
+import {getReposFromStr} from '../helpers';
+import {
+  fetchIssues
+} from '../redux/ducks/issue';
 import IssueList from './issue-list';
 import Issue from './issue';
-import Board from './board';
-
-/* Copy-pasta from MilestonesView */
 
 const KanbanColumn = React.createClass({
   render() {
@@ -25,20 +21,23 @@ const KanbanColumn = React.createClass({
           primaryRepoName={primaryRepoName}
           card={card}
           columnRegExp={columnRegExp}
-          />
+        />
       );
     });
 
     const heading = (
       <span className='user-title'>
-        <UserIcon/>
+        <PersonIcon/>
         {login}
       </span>
     );
 
     return (
-      <div key={login} className='kanban-board-column'>
-        <IssueList title={heading}>
+      <div className='kanban-board-column'>
+        <IssueList
+          title={heading}
+          login={login}
+        >
           {issueComponents}
         </IssueList>
       </div>
@@ -49,18 +48,28 @@ const KanbanColumn = React.createClass({
 
 
 const UsersView = React.createClass({
+  componentWillMount() {
+    const {repoInfos, dispatch} = this.props;
+    dispatch(fetchIssues(repoInfos));
+  },
   render() {
-    const {repoInfos, columnData, cards, columnRegExp} = this.props;
+    const {repoInfos, cards, columnRegExp} = this.props;
     const [{repoName}] = repoInfos; // primaryRepoName
 
-    let sortedCards = FilterStore.filterAndSort(cards, true);
+    const logins = new Set();
+    for (const card of this.props.cards) {
+      logins.add(card.issue.user.login);
+    }
+    const loginsArray = [];
+    for (const login of logins) {
+      loginsArray.push(login);
+    }
+    loginsArray.sort();
 
-    let kanbanColumnCount = 0; // Count the number of actual columns displayed
-
-    const kanbanColumns = _.map(columnData, (login) => {
+    const kanbanColumns = _.map(loginsArray, (login) => {
       // If we are filtering by a kanban column then only show that column
       // Otherwise show all columns
-      const columnCards = _.filter(sortedCards, (card) => {
+      const columnCards = _.filter(cards, (card) => {
         return card.issue.owner && card.issue.owner.login === login || card.issue.user.login === login;
       });
 
@@ -69,11 +78,9 @@ const UsersView = React.createClass({
       // isFilteringByColumn = label (the current column we are filtering on)
       // !isFilteringByColumn && (!getShowEmptyColumns || columnCards.length)
 
-      kanbanColumnCount++; // Count the number of actual columns displayed
-      /*HACK: Column should handle milestones */
       return (
         <KanbanColumn
-          key={kanbanColumnCount}
+          key={login}
           login={login}
           cards={columnCards}
           primaryRepoName={repoName}
@@ -90,52 +97,63 @@ const UsersView = React.createClass({
   }
 });
 
+export default connect((state, ownProps) => {
+  return {
+    repoInfos: getReposFromStr((ownProps.params || {}).repoStr || ''),
+    settings: state.settings,
+    cards: state.issues.cards,
+    columnRegExp: state.filter.columnRegExp,
+    milestones: state.issues.milestones,
+    filter: state.filter
+  };
+})(UsersView);
 
-const RepoKanbanShell = React.createClass({
-  displayName: 'RepoKanbanShell',
-  componentWillMount() {
-    // Needs to be called before `render()`
-    IssueStore.startPolling();
-  },
-  componentWillUnmount() {
-    IssueStore.stopPolling();
-  },
-  renderLoaded() {
-    const {repoInfos, columnRegExp} = getFilters().getState();
 
-    const columnDataPromise =
-      IssueStore.fetchIssues()
-      .then((cards) => {
-        const logins = new Set();
-        for (const card of cards) {
-          logins.add(card.issue.user.login);
-        }
-        const loginsArray = [];
-        for (const login of logins) {
-          loginsArray.push(login);
-        }
-        loginsArray.sort();
-        return loginsArray;
-      });
-
-    return (
-      <Board {...this.props}
-        repoInfos={repoInfos}
-        columnRegExp={columnRegExp}
-        type={UsersView}
-        columnDataPromise={columnDataPromise}
-      />
-    );
-  },
-  render() {
-
-    return (
-      <Loadable
-        promise={CurrentUserStore.fetchUser()}
-        renderLoaded={this.renderLoaded}
-      />
-    );
-  }
-});
-
-export default RepoKanbanShell;
+// const RepoKanbanShell = React.createClass({
+//   displayName: 'RepoKanbanShell',
+//   componentWillMount() {
+//     // Needs to be called before `render()`
+//     IssueStore.startPolling();
+//   },
+//   componentWillUnmount() {
+//     IssueStore.stopPolling();
+//   },
+//   renderLoaded() {
+//     const {repoInfos, columnRegExp} = getFilters().getState();
+//
+//     const columnDataPromise =
+//       IssueStore.fetchIssues()
+//       .then((cards) => {
+//         const logins = new Set();
+//         for (const card of cards) {
+//           logins.add(card.issue.user.login);
+//         }
+//         const loginsArray = [];
+//         for (const login of logins) {
+//           loginsArray.push(login);
+//         }
+//         loginsArray.sort();
+//         return loginsArray;
+//       });
+//
+//     return (
+//       <Board {...this.props}
+//         repoInfos={repoInfos}
+//         columnRegExp={columnRegExp}
+//         type={UsersView}
+//         columnDataPromise={columnDataPromise}
+//       />
+//     );
+//   },
+//   render() {
+//
+//     return (
+//       <Loadable
+//         promise={CurrentUserStore.fetchUser()}
+//         renderLoaded={this.renderLoaded}
+//       />
+//     );
+//   }
+// });
+//
+// export default RepoKanbanShell;
