@@ -1,14 +1,18 @@
 import React from 'react';
 import _ from 'underscore';
+import {connect} from 'react-redux';
 import {Link} from 'react-router';
 import * as BS from 'react-bootstrap';
 import classnames from 'classnames';
 import {BeakerIcon, SyncIcon, LockIcon, RepoForkedIcon, RepoIcon, PersonIcon, OrganizationIcon} from 'react-octicons';
-import history from '../history';
+
+import {
+  fetchRepositories,
+  fetchRepo
+} from '../redux/ducks/user';
+import {history} from '../redux/store';
 
 import {buildRoute} from '../route-utils';
-import Client from '../github-client';
-import CurrentUserStore from '../user-store';
 import AsyncButton from './async-button';
 import Time from './time';
 
@@ -127,7 +131,7 @@ const RepoGroup = React.createClass({
         selectedRepos[repoName] = true;
       }
       this.setState({selectedRepos});
-      // this.forceUpdate(); // since we are modifying hte object directly
+      // this.forceUpdate(); // since we are modifying the object directly
     };
   },
   goToBoard() {
@@ -157,7 +161,7 @@ const RepoGroup = React.createClass({
           repo={repo}
           isSelected={selectedRepos[repoName]}
           onSelect={this.toggleSelect(repoName)}
-          />
+        />
       );
     });
 
@@ -169,7 +173,7 @@ const RepoGroup = React.createClass({
     }
 
     let orgIcon;
-    if (CurrentUserStore.getUser() && CurrentUserStore.getUser().login === repoOwner) {
+    if (this.props.user && this.props.user.login === repoOwner) {
       orgIcon = (<PersonIcon className='org-icon'/>);
     } else {
       orgIcon = (<OrganizationIcon className='org-icon'/>);
@@ -196,9 +200,8 @@ const RepoGroup = React.createClass({
 });
 
 const Dashboard = React.createClass({
-  displayName: 'Dashboard',
   render() {
-    const {repos} = this.props;
+    const {repos, user} = this.props;
 
     const repoOwnersUpdatedAt = {};
     const reposByOwner = {};
@@ -229,6 +232,7 @@ const Dashboard = React.createClass({
       const groupRepos = reposByOwner[repoOwner];
       return (
         <RepoGroup
+          user={user}
           key={repoOwner}
           repoOwner={repoOwner}
           repos={groupRepos}
@@ -237,7 +241,7 @@ const Dashboard = React.createClass({
     });
 
     return (
-          <span>{repoOwnersNodes}</span>
+      <span>{repoOwnersNodes}</span>
     );
   }
 });
@@ -258,7 +262,7 @@ const CustomRepoModal = React.createClass({
   render() {
     const {customRepoName} = this.state;
     // Make sure the repo contains a '/'
-    const isInvalid = customRepoName && !/[^\/]\/[^\/]/.test(customRepoName);
+    const isInvalid = customRepoName && !/[^/]\/[^/]/.test(customRepoName);
     const isDisabled = !customRepoName || isInvalid;
     return (
       <BS.Modal {...this.props}>
@@ -280,10 +284,10 @@ const CustomRepoModal = React.createClass({
             disabled={isDisabled}
             bsStyle='primary'
             waitingText='Checking...'
-            action={() => Client.dbPromise().then(() => Client.getOcto().repos(customRepoName).fetchAll.bind(Client.getOcto()))}
+            action={() => this.props.dispatch(fetchRepo(customRepoName))}
             onResolved={() => this.goToBoard(customRepoName)}
             renderError={() => <span>Invalid Repo. Please try again.</span>}
-            >Show Board</AsyncButton>
+          >Show Board</AsyncButton>
         </BS.Modal.Footer>
       </BS.Modal>
     );
@@ -316,45 +320,32 @@ const ExamplesPanel = React.createClass({
             <RepoIcon className='repo-icon'/>
             Choose your own...
           </BS.ListGroupItem>
-          <CustomRepoModal show={showModal} container={this} onHide={close}/>
+          <CustomRepoModal show={showModal} container={this} onHide={close} dispatch={this.props.dispatch} />
         </BS.ListGroup>
       </BS.Panel>
     );
   }
 });
 
-
-let allMyReposHack = null;
-
 const DashboardShell = React.createClass({
   getInitialState() {
     return {repos: null};
   },
+  componentWillMount() {
+    if (this.props.user && !this.props.user.repositories) {
+      this.props.dispatch(fetchRepositories());
+    }
+  },
   render() {
-    let {repos} = this.state;
-
-    // HACK to not keep re-fetching the user's list of repos
-    if (repos) { allMyReposHack = repos; }
-    else { repos = allMyReposHack; }
+    let {user, ready, dispatch} = this.props;
 
     let myRepos;
 
-    if (repos) {
+    if (ready) {
       myRepos = (
-        <Dashboard repos={repos}/>
+        <Dashboard repos={(user || {}).repositories || []} user={user}/>
       );
     } else {
-      CurrentUserStore.fetchUser()
-      .then((currentUser) => {
-        if (currentUser) {
-          return Client.dbPromise().then(() => Client.getOcto().user.repos.fetchAll());
-        } else {
-          return []; // Anonymous has no repos
-        }
-      })
-      .then((allRepos) => this.setState({repos: allRepos}))
-      .catch(() => this.setState({repos: []}));
-      // TODO: Use Loadable component
       myRepos = (
         <span className='custom-loading is-loading'>
           <SyncIcon className='icon-spin'/>
@@ -368,7 +359,7 @@ const DashboardShell = React.createClass({
       <BS.Grid fluid className='dashboard' data-org-count={myRepos.length}>
         <BS.Row>
           <BS.Col md={6}>
-            <ExamplesPanel/>
+            <ExamplesPanel dispatch={dispatch} />
           </BS.Col>
           {myRepos}
         </BS.Row>
@@ -377,4 +368,9 @@ const DashboardShell = React.createClass({
   }
 });
 
-export default DashboardShell;
+export default connect((state) => {
+  return {
+    ready: state.user.ready,
+    user: state.user.info
+  };
+})(DashboardShell);
