@@ -1,20 +1,23 @@
 import React from 'react';
+import {connect} from 'react-redux';
 
-import Client from '../github-client';
-import IssueStore from '../issue-store';
+import {
+  selectors,
+  updateIssue
+} from '../redux/ducks/issue';
 
 import Etherpad from './etherpad';
 
 
 const EtherpadIssueShell = React.createClass({
   render() {
+    const {card, dispatch} = this.props;
     const {repoOwner, repoName, number} = this.props.params;
 
     const title = `Editing ${repoOwner}/${repoName}#${number}`;
     const padName = `issue_github.com_${repoOwner}_${repoName}_${number}`;
 
     const getBody = () => {
-      const card = IssueStore.issueNumberToCard(repoOwner, repoName, number);
       if (card.issue) {
         return card.issue.body;
       } else {
@@ -22,11 +25,9 @@ const EtherpadIssueShell = React.createClass({
       }
     };
     const saveBody = (text) => {
-      return Client.getOcto().repos(repoOwner, repoName).issues(number).update({body: text});
+      return dispatch(updateIssue(card, {body: text}));
     };
     const loadBody = () => {
-      const card = IssueStore.issueNumberToCard(repoOwner, repoName, number);
-
       // refetch the Issue body (esp if it hasn't been loaded yet)
       return card.fetchIssue().then(() => {
         return card.issue.body; // just in case someone uses this promise
@@ -39,4 +40,8 @@ const EtherpadIssueShell = React.createClass({
   }
 });
 
-export default EtherpadIssueShell;
+export default connect((state, ownProps) => {
+  return {
+    card: selectors.getCard(state.issues, ownProps.params)
+  };
+})(EtherpadIssueShell);
