@@ -1,20 +1,17 @@
 import React from 'react';
 import * as BS from 'react-bootstrap';
 import classnames from 'classnames';
-import {LockIcon, ListUnorderedIcon, GitPullRequestIcon, IssueOpenedIcon, IssueReopenedIcon, IssueClosedIcon} from 'react-octicons';
+import {ListUnorderedIcon, GitPullRequestIcon, IssueOpenedIcon, IssueReopenedIcon, IssueClosedIcon, QuestionIcon} from 'react-octicons';
 
 import {getFilters} from '../route-utils';
-import Database from '../database';
 
 import GithubFlavoredMarkdown from './gfm';
-import Loadable from './loadable';
 import ColoredIcon from './colored-icon';
 
 const IssueOrPullRequestBlurb = React.createClass({
   render() {
-    const {card, primaryRepoName, context, repo} = this.props;
+    const {card, primaryRepoName, context} = this.props;
     const {issue, repoOwner, repoName, number} = card;
-    const isPrivate = repo && repo.isPrivate;
 
     const multipleRepoName = primaryRepoName === repoName ? null : repoName;
     let blurbContext;
@@ -24,12 +21,6 @@ const IssueOrPullRequestBlurb = React.createClass({
       );
     }
 
-    let privateEl;
-    if (isPrivate) {
-      privateEl = (
-        <LockIcon className='private' title='Private Repository'/>
-      );
-    }
     if (issue) {
       const isPullRequest = card.isPullRequest() || issue.base; // use .base in case we are given the PR JSON (which does not contain labels)
       const popoverTitle = (
@@ -101,7 +92,7 @@ const IssueOrPullRequestBlurb = React.createClass({
           <a className='blurb-number-link'
             target='_blank'
             href={issue.htmlUrl}
-            >
+          >
             <span className='blurb-number'>{issue.number}</span>
           </a>
           <BS.OverlayTrigger
@@ -111,7 +102,6 @@ const IssueOrPullRequestBlurb = React.createClass({
             overlay={bodyPopover}>
             <span className='-just-for-overlay-trigger'>
               {icon}
-              {privateEl}
               <span className='blurb-secondary-repo'>{multipleRepoName}</span>
               {blurbContext}
             </span>
@@ -123,7 +113,6 @@ const IssueOrPullRequestBlurb = React.createClass({
       return (
         <span className='issue-blurb'>
           <span className='blurb-number'>{card.number}</span>
-          {privateEl}
           <span className='blurb-secondary-repo'>{multipleRepoName}</span>
           {blurbContext}
         </span>
@@ -134,20 +123,4 @@ const IssueOrPullRequestBlurb = React.createClass({
   }
 });
 
-
-const IssueOrPullRequestBlurbShell = React.createClass({
-  render() {
-    const {card, primaryRepoName, context} = this.props;
-    const {repoOwner, repoName} = card;
-    const promise = Database.getRepoOrNull(repoOwner, repoName);
-
-    return (
-      <Loadable
-        promise={promise}
-        renderLoaded={(repo) => (<IssueOrPullRequestBlurb repo={repo} card={card} primaryRepoName={primaryRepoName} context={context} />)}
-      />
-    );
-  }
-});
-
-export default IssueOrPullRequestBlurbShell;
+export default IssueOrPullRequestBlurb;
