@@ -25,13 +25,12 @@ const LabelBadge = React.createClass({
 
     if (KANBAN_LABEL.test(label.name)) {
       icon = (<ListUnorderedIcon/>);
-      name = label.name.replace(/^\d+\ -\ /, ' ');
+      name = label.name.replace(/^\d+ - /, ' ');
     } else {
       icon = (<TagIcon/>);
       name = label.name;
     }
     if (label.color && isLight(label.color)) {
-      className = className;
       className += ' ' + 'is-light';
     }
     if (extra) {
