@@ -1,4 +1,4 @@
-import d3 from 'd3';
+import d3 from 'd3'; // eslint-disable-line
 /**
  * From https://github.com/dk8996/Gantt-Chart
  * @author Dimitry Kudrayvtsev
@@ -21,7 +21,7 @@ export default function(milestoneCount) {
   let timeDomainMode = FIT_TIME_DOMAIN_MODE;// fixed or fit
   let taskTypes = [];
   let taskStatus = [];
-    // var height = document.body.clientHeight - margin.top - margin.bottom-5;
+  // var height = document.body.clientHeight - margin.top - margin.bottom-5;
   let height = milestoneCount * 50; // height of chart is based on # of milestones
   let width = document.body.clientWidth - margin.right - margin.left-5;
 
@@ -29,28 +29,28 @@ export default function(milestoneCount) {
 
   function keyFunction(d) {
     return d.startDate + d.taskName + d.endDate;
-  };
+  }
 
   function rectTransform(d) {
     return 'translate(' + x(d.startDate) + ',' + y(d.taskName) + ')';
-  };
+  }
   function rectTransformCompleted(d) {
     return 'translate(' + x(d.startDate) + ',' + y(d.taskName) + ')';
-  };
+  }
   function rectTransformRemaining(d) {
     return 'translate(' + (x(d.startDate) + d.percent * (x(d.endDate)-x(d.startDate))) + ',' + y(d.taskName) + ')';
-  };
+  }
   function rectTransformSegment({task, segment, total, prev}) {
     const percent = prev / total;
     return 'translate(' + (x(task.startDate) + percent * (x(task.endDate)-x(task.startDate))) + ',' + y(task.taskName) + ')';
-  };
+  }
 
   let x = d3.time.scale().domain([ timeDomainStart, timeDomainEnd ]).range([ 0, width ]).clamp(true);
 
   let y = d3.scale.ordinal().domain(taskTypes).rangeRoundBands([ 0, height - margin.top - margin.bottom ], .1);
 
   let xAxis = d3.svg.axis().scale(x).orient('bottom').tickFormat(d3.time.format(tickFormat)).tickSubdivide(true)
-	    .tickSize(8).tickPadding(8);
+    .tickSize(8).tickPadding(8);
 
   let yAxis = d3.svg.axis().scale(y).orient('right').tickSize(0);
 
@@ -70,31 +70,31 @@ export default function(milestoneCount) {
       });
       timeDomainStart = tasks[0].startDate;
     }
-  };
+  }
 
   function initAxis() {
     x = d3.time.scale().domain([ timeDomainStart, timeDomainEnd ]).range([ 0, width ]).clamp(true);
     y = d3.scale.ordinal().domain(taskTypes).rangeRoundBands([ 0, height - margin.top - margin.bottom ], .1);
     xAxis = d3.svg.axis().scale(x).orient('bottom').tickFormat(d3.time.format(tickFormat)).tickSubdivide(true)
-		  .tickSize(8).tickPadding(8);
+      .tickSize(8).tickPadding(8);
 
     yAxis = d3.svg.axis().scale(y).orient('right').tickSize(0);
-  };
+  }
 
   function gantt(tasks) {
     initTimeDomain(tasks);
     initAxis();
 
     const svg = d3.select(selector)
-	    .append('svg')
-	    .attr('class', 'chart')
-	    .attr('width', width + margin.left + margin.right)
-    	.attr('height', height + margin.top + margin.bottom)
-    	.append('g')
+      .append('svg')
+      .attr('class', 'chart')
+      .attr('width', width + margin.left + margin.right)
+      .attr('height', height + margin.top + margin.bottom)
+      .append('g')
       .attr('class', 'gantt-chart')
-    	.attr('width', width + margin.left + margin.right)
-    	.attr('height', height + margin.top + margin.bottom)
-    	.attr('transform', 'translate(' + margin.left + ', ' + margin.top + ')');
+      .attr('width', width + margin.left + margin.right)
+      .attr('height', height + margin.top + margin.bottom)
+      .attr('transform', 'translate(' + margin.left + ', ' + margin.top + ')');
 
     const PHIL = svg.selectAll('.chart');
 
@@ -108,15 +108,15 @@ export default function(milestoneCount) {
       .style({stroke:'rgb(0,0,255)', 'stroke-width':2});
 
     PHIL.selectAll('g.milestone-bars').data(tasks, keyFunction).enter()
-  	  .append('g')
-  	//  .attr("rx", 5)
+      .append('g')
+    //  .attr("rx", 5)
     //        .attr("ry", 5)
-  	  .attr('class', 'milestone-bars')
-  	//  .attr("y", 0)
-  	//  .attr("transform", rectTransformCompleted)
-  	//  .attr("height", function(d) { return y.rangeBand(); })
-  	//  .attr("width", function(d) {
-  	//      return d.percent * (x(d.endDate) - x(d.startDate));
+      .attr('class', 'milestone-bars')
+    //  .attr("y", 0)
+    //  .attr("transform", rectTransformCompleted)
+    //  .attr("height", function(d) { return y.rangeBand(); })
+    //  .attr("width", function(d) {
+    //      return d.percent * (x(d.endDate) - x(d.startDate));
     //    })
       .selectAll('rect').data((task) => {task.segments.map((segment) => ({task, segment}));}).enter()
       .append('rect')
@@ -130,33 +130,33 @@ export default function(milestoneCount) {
       });
 
 
-   // PHIL: Copy-pasta to show the 2nd half of the milestone (unfinished)
- //   PHIL.selectAll('rect.milestone-remaining').data(tasks, keyFunction).enter()
- //
- // 	 .append("rect")
- // // 	 .attr("rx", 5)
- //  //         .attr("ry", 5)
- // 	 .attr("class", function(d){
- // 	     return 'milestone-remaining ' + d.status;
- // 	     })
- // 	 .attr("y", 0)
- // 	 .attr("transform", rectTransformRemaining)
- // 	 .attr("height", function(d) { return y.rangeBand(); })
- // 	 .attr("width", function(d) {
- // 	     return (1-d.percent) * (x(d.endDate) - x(d.startDate));
- // 	     });
+    // PHIL: Copy-pasta to show the 2nd half of the milestone (unfinished)
+    //   PHIL.selectAll('rect.milestone-remaining').data(tasks, keyFunction).enter()
+    //
+    // 	 .append("rect")
+    // // 	 .attr("rx", 5)
+    //  //         .attr("ry", 5)
+    // 	 .attr("class", function(d){
+    // 	     return 'milestone-remaining ' + d.status;
+    // 	     })
+    // 	 .attr("y", 0)
+    // 	 .attr("transform", rectTransformRemaining)
+    // 	 .attr("height", function(d) { return y.rangeBand(); })
+    // 	 .attr("width", function(d) {
+    // 	     return (1-d.percent) * (x(d.endDate) - x(d.startDate));
+    // 	     });
 
 
     svg.append('g')
-  	  .attr('class', 'x axis')
-  	  .attr('transform', 'translate(0, ' + (height - margin.top - margin.bottom) + ')')
-  	  .transition()
-  	  .call(xAxis);
+      .attr('class', 'x axis')
+      .attr('transform', 'translate(0, ' + (height - margin.top - margin.bottom) + ')')
+      .transition()
+      .call(xAxis);
 
     svg.append('g').attr('class', 'y axis').attr('transform', 'translate(-' + margin.left + ',0)').transition().call(yAxis);
 
     return gantt;
-  };
+  }
 
   gantt.redraw = function(tasks) {
 
@@ -171,12 +171,12 @@ export default function(milestoneCount) {
     rect.enter()
       .append('g')
       .attr('class', 'milestone-bars')
-       // 	 .attr("y", 0)
-       // 	 .attr("transform", rectTransformCompleted)
-       // 	 .attr("height", function(d) { return y.rangeBand(); })
-       // 	 .attr("width", function(d) {
-       // 	     return d.percent * (x(d.endDate) - x(d.startDate));
-        //     })
+    // 	 .attr("y", 0)
+    // 	 .attr("transform", rectTransformCompleted)
+    // 	 .attr("height", function(d) { return y.rangeBand(); })
+    // 	 .attr("width", function(d) {
+    // 	     return d.percent * (x(d.endDate) - x(d.startDate));
+    //     })
       .selectAll('rect').data((task) => {
         const total = task.segments.reduce((acc, segment) => acc + segment.count, 0);
         let prev = 0;
@@ -187,11 +187,11 @@ export default function(milestoneCount) {
         });
         return segments;
       }).enter()
-     .append('rect')
-          //  .attr('class', 'milestone-segment')
+      .append('rect')
+    //  .attr('class', 'milestone-segment')
       .style('fill', ({task, segment}) => `#${segment.color}`)
-          // .attr('class', ({task, segment}) => {return {fill: segment.color}; })
-           // .transform()
+    // .attr('class', ({task, segment}) => {return {fill: segment.color}; })
+    // .transform()
       .attr('transform', rectTransformSegment)
       .attr('height', (d) => y.rangeBand())
       .attr('width', ({task, total, segment}) => {
@@ -200,54 +200,54 @@ export default function(milestoneCount) {
       });
 
 
-  //       rect.enter()
-  //        .append("rect")
-	//  .attr("class", function(d){
-	//      return 'milestone-bars ' +d.status;
-	//      })
-	//  .transition()
-	//  .attr("y", 0)
-	//  .attr("transform", rectTransformCompleted)
-	//  .attr("height", function(d) { return y.rangeBand(); })
-	//  .attr("width", function(d) {
-	//      return d.percent*(x(d.endDate) - x(d.startDate));
-	//      });
-   //
-  //       rect.transition()
-  //         .attr("transform", rectTransform)
-	//  .attr("height", function(d) { return y.rangeBand(); })
-	//  .attr("width", function(d) {
-	//      return d.percent*(x(d.endDate) - x(d.startDate));
-	//      });
+    //       rect.enter()
+    //        .append("rect")
+    //  .attr("class", function(d){
+    //      return 'milestone-bars ' +d.status;
+    //      })
+    //  .transition()
+    //  .attr("y", 0)
+    //  .attr("transform", rectTransformCompleted)
+    //  .attr("height", function(d) { return y.rangeBand(); })
+    //  .attr("width", function(d) {
+    //      return d.percent*(x(d.endDate) - x(d.startDate));
+    //      });
+    //
+    //       rect.transition()
+    //         .attr("transform", rectTransform)
+    //  .attr("height", function(d) { return y.rangeBand(); })
+    //  .attr("width", function(d) {
+    //      return d.percent*(x(d.endDate) - x(d.startDate));
+    //      });
 
-// 	rect.exit().remove();
-//
-//   //PHIL Copy-pasta
-//   var rect = ganttChartGroup.selectAll("rect.milestone-remaining").data(tasks, keyFunction);
-//
-//   rect.enter()
-//    .append("rect")
-//    .attr("rx", 5)
-//    .attr("ry", 5)
-// .attr("class", function(d){
-//  return 'milestone-remaining ' +d.status;
-//  })
-// .transition()
-// .attr("y", 0)
-// .attr("transform", rectTransformRemaining)
-// .attr("height", function(d) { return y.rangeBand(); })
-// .attr("width", function(d) {
-//  return (1-d.percent)*(x(d.endDate) - x(d.startDate));
-//  });
-//
-//   rect.transition()
-//     .attr("transform", rectTransformRemaining)
-// .attr("height", function(d) { return y.rangeBand(); })
-// .attr("width", function(d) {
-//  return (1-d.percent)*(x(d.endDate) - x(d.startDate));
-//  });
-//
-// rect.exit().remove();
+    // 	rect.exit().remove();
+    //
+    //   //PHIL Copy-pasta
+    //   var rect = ganttChartGroup.selectAll("rect.milestone-remaining").data(tasks, keyFunction);
+    //
+    //   rect.enter()
+    //    .append("rect")
+    //    .attr("rx", 5)
+    //    .attr("ry", 5)
+    // .attr("class", function(d){
+    //  return 'milestone-remaining ' +d.status;
+    //  })
+    // .transition()
+    // .attr("y", 0)
+    // .attr("transform", rectTransformRemaining)
+    // .attr("height", function(d) { return y.rangeBand(); })
+    // .attr("width", function(d) {
+    //  return (1-d.percent)*(x(d.endDate) - x(d.startDate));
+    //  });
+    //
+    //   rect.transition()
+    //     .attr("transform", rectTransformRemaining)
+    // .attr("height", function(d) { return y.rangeBand(); })
+    // .attr("width", function(d) {
+    //  return (1-d.percent)*(x(d.endDate) - x(d.startDate));
+    //  });
+    //
+    // rect.exit().remove();
 
 
     svg.select('.x').transition().call(xAxis);
@@ -325,4 +325,4 @@ export default function(milestoneCount) {
   };
 
   return gantt;
-};
+}
