@@ -1,9 +1,23 @@
 import 'babel-polyfill';
+import React from 'react';
+import ReactDOM from 'react-dom';
+import {Router} from 'react-router';
+import {Provider} from 'react-redux';
 import Recordo from 'recordo';
-import startRouter from './router';
-import CurrentUserStore from './user-store';
+import {store, history} from './redux/store';
+import {fetchEmojis} from './redux/ducks/emojis';
+import routes from './router';
 
 Recordo.initialize();
-startRouter();
 
-CurrentUserStore.fetchEmojis();
+// Use hashHistory because there is a <MenuItem> in <App> that explicitly tacks on a "#"
+// Also, gh-pages does not support arbitrary URLs so the repos need to be in the hash
+const root = (
+  <Provider store={store}>
+    <Router history={history} routes={routes} />
+  </Provider>
+);
+
+ReactDOM.render(root, document.getElementById('container'));
+
+store.dispatch(fetchEmojis());
