@@ -1,3 +1,1 @@
-/*eslint-disable no-unused-vars */
-import styles from './app.less';
-/*eslint-enable no-unused-vars */
+import './app.less';
